+++ Source/Coinbase.Pro/CoinbaseProClient.Withdrawals.cs
@@ -37,8 +37,14 @@ namespace Coinbase.Pro
       /// <summary>
       /// Get a list of withdrawals from the profile of the API key, in descending order by created time.
       /// </summary>
-      Task<PagedResponse<Withdrawal>> ListWithdrawals(string type = "withdraw", string profileId = null,
+      Task<PagedResponse<Withdrawal>> ListWithdrawals(string profileId = null,
          int? limit = null, long? before = null, long? after = null,
+         CancellationToken cancellationToken = default);
+
+      /// <summary>
+      /// Get information on a single withdrawal.
+      /// </summary>
+      Task<Withdrawal> GetWithdrawal(string transferId = null,
          CancellationToken cancellationToken = default);
    }
 
@@ -101,17 +107,25 @@ namespace Coinbase.Pro
             .ReceiveJson<CryptoWithdraw>();
       }
 
-      Task<PagedResponse<Withdrawal>> IWithdrawalsEndpoint.ListWithdrawals(
-         string type, string profileId,
+      Task<PagedResponse<Withdrawal>> IWithdrawalsEndpoint.ListWithdrawals(string profileId,
          int? limit, long? before, long? after,
          CancellationToken cancellationToken)
       {
          return this.TransfersEndpoint
             .WithClient(this)
-            .SetQueryParam("type", type)
+            .SetQueryParam("type", "withdraw")
             .SetQueryParam("profile_id", profileId)
             .AsPagedRequest(limit, before, after)
             .GetPagedJsonAsync<Withdrawal>(cancellationToken);
+      }
+
+      Task<Withdrawal> IWithdrawalsEndpoint.GetWithdrawal(string transferId,
+         CancellationToken cancellationToken)
+      {
+         return this.TransfersEndpoint
+            .WithClient(this)
+            .SetQueryParam("transfer_id", transferId)
+            .GetJsonAsync<Withdrawal>(cancellationToken);
       }
    }
 }